Changed checks for loglinear option.
When the loglinear approximation is used, Dynare does not actually compute the
loglinear approximation but rewrite the model in percentage deviation to the
steady state (without the log approximation). Consequently, there is no reason
to reject parameterizations where the steady state is negative.

Maybe we should change the behaviour of this option, by effectively computing
the loglinear approximation (rewritting the model's equations in terms of the
log of the original variables). An immediate advantage, would be that we could
easily expand that to higher order approximations. The cuurrent behaviour would
then be triggered by a new option (percentagedeviations?).